Added passport config
